docs: update README and consolidate documentation
- Replace README with accurate, verified metrics only
- Add streamlined docs: QUICKSTART, TESTING, METRICS, RESUME, GITHUB_SETUP
- Remove false performance claims (50K e/s, 95% precision)
- Highlight verified metrics: sub-30ms P99, 100% Kafka delivery
- Professional formatting with tables and architecture diagram

+++ services/ingestion/consumer.go
@@ -297,7 +297,7 @@ func (sw *StorageWriter) writeBatch(events []*models.Event) {
 	log.Info().
 		Int("batch_size", len(events)).
 		Int("success_count", successCount).
-		Float64("latency_ms", time.Since(startTime).Milliseconds()).
+		Float64("latency_ms", float64(time.Since(startTime).Milliseconds())).
 		Msg("Batch written to database")
 }
